Rebuild __init__ of ContestPopulation
The ContestPopulation was missing all new features of the Population,
i.e. logging and serialization. I've added these, I've updated the
docstring and I've properly typed the arguments.

The typing results in a little problem, as the ContestPopulation expects
an eval_function with a different type than the original Population.
PyCharm detects this and complains. I see no way to go around this other
than relaxing the typing of the eval_function in Population, which I do
not want to do.