Remove generic combiners (#143)
The generic combiners have a signature that does not match our expected
signature for combiners. While we expect a combiner to accept a number
of chromosomes, these combiners except individuals.

As I think that a combiner should solely combine two chromosomes, and
all logic related to the fitness of individuals selected for combining
should be handled by the picker, I've removed them.